stock-data-extraction/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
Flask==2.0.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
polygon-api-client==0.2.11
requests==2.26.0
urllib3==1.26.7
websocket-client==1.2.1
websockets==10.1
Werkzeug==2.0.2
